Conflicts:
	Files/Classes/PreferenceHandler.m
	Files/DLS Import.c
	Files/FFT-Dialog.c
	Files/Functions.c
	Files/Help.c
	Files/InternetMapping.c
	Files/MIDI-GM.c
	Files/Main.c
	Files/wds_views/Spectrum.c

+++ Files/wds_editors/StaffEditor.c
@@ -330,7 +330,7 @@ void DoNullStaff(void)
 		
 		/**/
 		
-		NumToString((long) CurrentPat, String);
+		NumToString(CurrentPat, String);
 		pStrcpy(aStr, "\p");
 		if (CurrentPat < 10) pStrcat(aStr, "\p0");
 		if (CurrentPat < 100) pStrcat(aStr, "\p0");
